+++ src/sql/ggplot/geom/geom_histogram.py
@@ -18,12 +18,16 @@ class geom_histogram(geom):
 
     cmap : str, default 'viridis
         Apply a color map to the stacked graph
+
+    breaks : list
+        Divide bins with custom intervals
     """
 
-    def __init__(self, bins=None, fill=None, cmap=None, **kwargs):
+    def __init__(self, bins=None, fill=None, cmap=None, breaks=None, **kwargs):
         self.bins = bins
         self.fill = fill
         self.cmap = cmap
+        self.breaks = breaks
         super().__init__(**kwargs)
 
     @telemetry.log_call("ggplot-histogram")
@@ -40,5 +44,6 @@ class geom_histogram(geom):
             edgecolor=gg.mapping.color,
             facet=facet,
             ax=ax or gg.axs[0],
+            breaks=self.breaks,
         )
         return gg