+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 0.9.1dev
+* [Feature] Added `--breaks/-B` to ggplot histogram for specifying breaks (#719)
 * [Fix] Fix boxplot for duckdb native ([#728](https://github.com/ploomber/jupysql/issues/728))
 * [Doc] Add Redshift tutorial
 * [Feature] Adds Redshift support for `%sqlplot boxplot`